agentcodebase/pyproject.toml
[project]
name = "agentcodebase"
version = "0.1.1"
description = "Highly Reusable Utility Modules and Codebase for Agentic Scientific Research"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "matplotlib>=3.10.8",
    "openai>=2.16.0",
    "paramiko>=4.0.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
